+++ src/constants.js
@@ -3,3 +3,4 @@ export const UAS = 'uas';
 export const ASSIGNED_FILTER_KEY = 'uas.Assigned';
 export const UNASSIGNED_FILTER_KEY = 'uas.Unassigned';
 export const ASSIGNED = 'Assigned';
+export const UNASSIGNED = 'Unassigned';